+++ supportedBackends/google/batch/src/main/scala/cromwell/backend/google/batch/models/GcpBatchConfigurationAttributes.scala
@@ -25,6 +25,7 @@ import cromwell.backend.google.batch.models.GcpBatchConfigurationAttributes.{
 import cromwell.backend.google.batch.util.GcpBatchReferenceFilesMappingOperations
 import cromwell.cloudsupport.gcp.GoogleConfiguration
 import cromwell.cloudsupport.gcp.auth.GoogleAuthMode
+import cromwell.docker.DockerMirroring
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.numeric.Positive
 import eu.timepit.refined.{refineMV, refineV}
@@ -49,6 +50,7 @@ case class GcpBatchConfigurationAttributes(
   cacheHitDuplicationStrategy: BatchCacheHitDuplicationStrategy,
   requestWorkers: Int Refined Positive,
   batchTimeout: FiniteDuration,
+  dockerMirroringOpt: Option[DockerMirroring],
   logFlushPeriod: Option[FiniteDuration],
   gcsTransferConfiguration: GcsTransferConfiguration,
   virtualPrivateCloudConfiguration: VirtualPrivateCloudConfiguration,
@@ -111,6 +113,8 @@ object GcpBatchConfigurationAttributes extends GcpBatchReferenceFilesMappingOper
     "request-workers",
     "batch-timeout",
     "max-transient-error-retries",
+    "docker-mirror.dockerhub.enabled",
+    "docker-mirror.dockerhub.address",
     "batch-requests.timeouts.read",
     "batch-requests.timeouts.connect",
     "default-runtime-attributes.bootDiskSizeGb",
@@ -249,6 +253,8 @@ object GcpBatchConfigurationAttributes extends GcpBatchReferenceFilesMappingOper
 
     val batchTimeout: FiniteDuration = backendConfig.getOrElse("batch-timeout", 7.days)
 
+    val dockerMirroring: Option[DockerMirroring] = DockerMirroring.fromConfig(backendConfig)
+
     val logFlushPeriod: Option[FiniteDuration] = backendConfig.as[Option[FiniteDuration]]("log-flush-period") match {
       case Some(duration) if duration.isFinite => Option(duration)
       // "Inf" disables upload
@@ -340,6 +346,7 @@ object GcpBatchConfigurationAttributes extends GcpBatchReferenceFilesMappingOper
           cacheHitDuplicationStrategy = cacheHitDuplicationStrategy,
           requestWorkers = requestWorkers,
           batchTimeout = batchTimeout,
+          dockerMirroringOpt = dockerMirroring,
           logFlushPeriod = logFlushPeriod,
           gcsTransferConfiguration = gcsTransferConfiguration,
           virtualPrivateCloudConfiguration = virtualPrivateCloudConfiguration,